requirements.txt
numpy==1.26.4
pandas==2.2.3
zarr==2.18.3
timm==0.9.8
transformers==4.51.3
huggingface-hub==0.31.1
pillow==11.2.1
opencv-python==4.11.0.86
tifffile==2025.5.10
scikit-learn==1.5.2
scikit-survival==0.23.1
xgboost==3.0.2
matplotlib==3.10.3
seaborn==0.13.2
scipy==1.15.3
networkx==3.4.2
tqdm==4.67.1
lifelines==0.30.0

# Additional dependencies for MUSK
einops==0.8.1
safetensors==0.5.3
fairscale==0.4.13
